feat(cca-template): Upgrade react-router-dom to 6.4.5
Split App component to AppRouter & AppLayout
