+++ packages/cozy-scripts/template/package.json
@@ -50,6 +50,6 @@
     "cozy-ui": "^78.1.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-router-dom": "5.2.0"
+    "react-router-dom": "6.4.5"
   }
 }
